+++ src/ScoreFilter/SunScoreFilter.cs
@@ -8,12 +8,12 @@ public sealed class SunScoreFilter : IScoreFilter
 {
     public Func<CommonRecord, bool> GetFilter(string? condition) => x =>
     {
-        float sumScore = (x.Chart.Notes.Tap + x.Chart.Notes.Touch + (x.Chart.Notes.Hold * 2) +
+        decimal sumScore = (x.Chart.Notes.Tap + x.Chart.Notes.Touch + (x.Chart.Notes.Hold * 2) +
                           (x.Chart.Notes.Slide * 3) + (x.Chart.Notes.Break * 5)) * 5;
-        float minScore = Math.Min((1 - ((sumScore - 1) / sumScore)) * 100,
-            x.Chart.Notes.Break > 0 ? 1f / x.Chart.Notes.Break / 2 : 101);
-        float maxAcc = ConstantMap.GetRatingMinAchievement(x.Rank + 1);
-        float minAcc = maxAcc - minScore;
+        decimal minScore = Math.Min((1 - ((sumScore - 1) / sumScore)) * 100,
+            x.Chart.Notes.Break > 0 ? 1m / x.Chart.Notes.Break / 2 : 101);
+        decimal maxAcc = ConstantMap.GetRatingMinAchievement(x.Rank + 1);
+        decimal minAcc = maxAcc - minScore;
         return x.Achievements >= minAcc && x.Achievements < maxAcc;
     };
 }
