+++ src/ScoreProcesser/StdDevScoreProcesser.cs
@@ -12,8 +12,8 @@ public sealed class StdDevScoreProcesser : IScoreProcesser
     {
         ParallelQuery<CommonRecord> rankedRecords = records.AsParallel().Select(record =>
             {
-                double stdDev = 0;
-                double fitLevel = record.Chart.LevelValue;
+                decimal stdDev = 0;
+                decimal fitLevel = record.Chart.LevelValue;
                 if (Status.Shared.TryGetChartState(record.Chart.Song.Id, (int)record.Chart.Difficulty - 1,
                         out ChartState? chartState))
                 {
@@ -21,8 +21,8 @@ public sealed class StdDevScoreProcesser : IScoreProcesser
                     fitLevel = chartState.FitLevel;
                 }
 
-                record.ExtraInfo = (float)stdDev;
-                double score = record.DXRating * (1 + (stdDev / 10)) * (1 + (fitLevel / (fitLevel > 0 ? 10 : 1)));
+                record.ExtraInfo = stdDev;
+                decimal score = record.DXRating * (1 + (stdDev / 10)) * (1 + (fitLevel / (fitLevel > 0 ? 10 : 1)));
                 return (Record: record, Score: score);
             }).OrderByDescending(x => x.Score).ThenByDescending(x => x.Record.Chart.LevelValue)
             .ThenByDescending(x => x.Record.Achievements).Select(x => x.Record);
